refact(backend): rename GenerateDTOs to RequestFull DTOs and update references
