+++ backend/my-way/src/main/java/br/ufrn/myway/Controller/RoadmapController/RoadmapController.java
@@ -2,7 +2,7 @@ package br.ufrn.myway.Controller.RoadmapController;
 
 import java.util.List;
 
-import br.ufrn.myway.Model.DTO.GenerateRoadmapDTO;
+import br.ufrn.myway.Model.DTO.Request.RequestFullRoadmapDTO;
 import br.ufrn.myway.Model.DTO.Response.ResponseRoadmapDTO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -41,7 +41,7 @@ public class RoadmapController {
     }
 
     @PostMapping("/save/{userId}")
-    public ResponseEntity<ResponseRoadmapDTO> save(@RequestBody GenerateRoadmapDTO roadmapDTO, @PathVariable Long userId){
+    public ResponseEntity<ResponseRoadmapDTO> save(@RequestBody RequestFullRoadmapDTO roadmapDTO, @PathVariable Long userId){
         return ResponseEntity.status(HttpStatus.CREATED).body(
                 roadmapMapper.toResponse(
                         roadmapService.save(
