--- backend/my-way/src/main/java/br/ufrn/myway/Model/DTO/GenerateGoalDTO.java
@@ -1,11 +1,0 @@
-package br.ufrn.myway.Model.DTO;
-
-import java.util.List;
-
-public record GenerateGoalDTO(
-        String name,
-        String description,
-        Long roadmapIndex,
-        List<GenerateStudyTopicDTO> studyTopics
-) {
-}
