--- backend/my-way/src/main/java/br/ufrn/myway/Model/DTO/GenerateStudyTopicDTO.java
@@ -1,8 +1,0 @@
-package br.ufrn.myway.Model.DTO;
-
-public record GenerateStudyTopicDTO(
-        String name,
-        String description,
-        boolean isDone
-) {
-}
