+++ backend/my-way/src/main/java/br/ufrn/myway/Model/DTO/Request/RequestFullGoalDTO.java
@@ -1,0 +1,11 @@
+package br.ufrn.myway.Model.DTO.Request;
+
+import java.util.List;
+
+public record RequestFullGoalDTO(
+        String name,
+        String description,
+        Long roadmapIndex,
+        List<RequestFullStudyTopicDTO> studyTopics
+) {
+}
