+++ backend/my-way/src/main/java/br/ufrn/myway/Model/DTO/Request/RequestFullStudyTopicDTO.java
@@ -1,0 +1,8 @@
+package br.ufrn.myway.Model.DTO.Request;
+
+public record RequestFullStudyTopicDTO(
+        String name,
+        String description,
+        boolean isDone
+) {
+}
